pyproject.toml
[project]
name = "shedding-hub"
version = "0.1.0"
readme = "README.md"
description = "Data and statistical models for biomarker shedding."
dependencies = [
    "pyaml",
    "requests",
]

[tool.setuptools.packages]
find = {}
